Review of ImportAup.cpp ...
... incidental to a review of CopySamples in it.  This code is used only when
importing an old-style .aup project file into the new project format.

Fixed some RAII for file handle.

Rewrote every call to IsGoodInt or IsGoodInt64 with a narrowly scoped temporary
variable.

Use IsGoodInt64 only for total track lengths and positions; but use IsGoodInt
for block and blockfile sizes, which are not supposed to be huge but are
supposed to fit in memory buffers.