Issue 871: Let Audacity compile with unmodified portaudio sources...
... This may allow upgrade to the latest version of portaudio, which is
desirable for fixes and improvements, including the fixing of build warnings on
Mac about use of deprecated SDK functions.

The function PaUtil_GetTime() cannot be called.

To make a conservative fix that doesn't affect the fine tuning of timing,
replicate the implementation of that function directly in AudioIO.cpp, with
all necessary conditional compilation branches.

Don't attempt a fix with std::chrono, although that would be more nicely
platform independent.

Testing should confirm that MIDI playback doesn't suffer any jittery timing,
such as it did especially on Linux with ALSA, when the playback feature was
under development in 2017.