Specify narrow format where needed in all Append() and Set() calls...
... these are in: imports, draw tool, recording, mix-and-render, and these
effects: Reverse, Repair, Invert, and TruncateSilence.

Explicit effective format is given for all imports, though maybe redundantly
because it is often the same as the appended buffer's format.

Note that Invert will not dither, so inverting a track, mixing with the
original as exported and reimported, and amplifying, will be a way to test for
the presence of dither.