Simplify EffectManager::GetEffect...
... And remove EffectDefinitionInterface::IsLegacy, which was false for newly
instantiated plugins exactly when they did not inherit class Effect.  But the
dynamic_cast to Effect was done anyway in the one place where it was used.

Also remove unnecessary analogous code for AudacityCommand objects.

(Commonalities between Effects and Commands should perhaps be expressed in a
common base class in the future, but were assumed to be greater than they
really are when certain Command handling code was written.)