Update BatchCommands.cpp
Removed a call to stop playing the current project when a batch call is sent to Audacity. There are reasons to leave playback (especially looping playback) running while batch commands are being issued. Stop can be a part of someone's batch commands if they need to stop playback of the project.