Issue 1389: Autoduck can miss sections below threshold
Problem: the running average rmsSum can be higher than it should be due to rounding errors. This can lead to sections of the audio which should be below the threshold, being measured to be above the threshold.

Fix: make rmsSum a double, rather than a float, to greatly reduce rounding errors.