ProjectFileIORegistry is generated from a class template...
... And macros to help generate the one necessary function in it with linkage,
so that it works on Windows builds too.

Cause that function to be generated in Project.cpp.

Rename the files defining the template and its base class.

Now the base classes make no mention of class AudacityProject, not even as an
incomplete type.

Will this generality be needed?  Yes!  To merge Leland's experimental stackable
effects branch, but also to break some of the dependency cycles in it, this
template will be reused with Host = WaveTrack