AudacityMessageBox button presses can now be replayed in journal...
... the dialog won't actually reappear but the same value will be returned to
the program.

These message boxes are special native ones that don't use the wxWidgets event
framework.  Therefore the system for journalling of mouse button presses does
not work with them.  Treat them specially.

A generalized facility added to namespace Journal in case this pattern needs to
be repeated for other actions.