Fix journal of return key invoking default dialog button on Mac...
... and some other cases.

Wherever Audacity simulates button press events, set the event object to be the
button, so that WindowEventSerialization() in JournalEvents.cpp can determine
a pathname for the window associated with the event.

The change in KeyboardCapture is enough for the default buttons of dialogs. The
other changes are necessary but not sufficient for fixing journalling of some of
Audacity's custom widgets, on any operating system.