Functions for end of track processing...
... SpectrumTransformer is constructed with a flag for trailing padding, as
also for leading; and TrackSpectrumTransformer is responsible for allocating
the output track.

Look closely and you may see a resequencing of steps in the finishing procedure
-- but those steps were mutually exclusive possibilities in practice:  some
only for the noise profiling pass, in which there is no output track and no
trailing padding; others for the processing pass, in which, the opposites.