Restore missing localizations on Linux
The INSTALL_PREFIX variable in src/audacity_config.h.in was being
passed into the program as an empty string. The file is configured
in the top-level CMakeLists.txt but INSTALL_PREFIX is defined in
src/CMakeLists.txt. CMake variables are not passed to the parent scope
unless explicitly requested.

Partially fix #1382. Allows Audacity to find the translations when they
are at the default location specified by INSTALL_PREFIX, but not if
they have been moved elsewhere. This works for distro builds but it's
not enough to get the AppImage working again.