Issue 1681: Position of New Label dialog is inconvenient
If the "use dialog for name of new label" check box is checked in the track behaviour page of preferences, then when adding a label at either selection or playback position, a dialog is used for entering the name of the new label. For convenience the horizontal position of the dialog should be such that the initial horizontal position of the cursor in the edit box is roughly the same as the horizontal position of the cursor or playback position respectively.

This behaviour was broken by commit 8f8ec41.
In this commit in LabelTrackView::DialogForLabelName(), the two arguments passed to viewInfo.TimeToPosition() are in wrong way round.

The fix: swap the order of these arguments.