Issue 1681: Position of New Label dialog is inconvenient
If the "use dialog for name of new label" check box is checked in the track behaviour page of preferences, then when adding a label at either selection or playback position, a dialog is used for entering the name of the new label. For convenience the horizontal position of the dialog should be such that the initial horizontal position of the cursor in the edit box is roughly the same as the horizontal position of the cursor or playback position respectively.

This behaviour was broken by commit 8f8ec41.
In LabelTrackView::DialogForLabelName(), there was the change from:
   position.x += viewInfo.GetLabelWidth()
      + std::max(0, static_cast<int>(viewInfo.TimeToPosition(region.t0())))
      -40;

to:
   position.x +=
      + std::max(0, static_cast<int>(viewInfo.TimeToPosition(
         viewInfo.GetLeftOffset(), region.t0())))
      - 39;

The change that was causing the most problem is that the arguments passed to viewInfo.TimeToPosition() are the wrong way round. However in addition the original code guarded against the position of the cursor being negative - in that case it was treated as zero (this prevented the dialog from going off the left of Audacity's main window).

The fix: revert to the original structure of the statement.