Increased max FFT size in Plot Spectrum to enable higher resolution
The maximum FFT size in Plot Spectrum has been increased by a factor of 2 to 131072. This change has been made to allow a higher resolution FFT to be computed. This can be useful when assessing low frequency content of audio waveforms in music. This gives change gives much better than 1Hz resolution at a sampling frequency of 48kHz. At a sample rate of 192kHz, a 1.47Hz resolution is obtained, which is deemed acceptable.

Also, some minor typos have been corrected, and some of the code formatting adjusted for consistency.