Update FreqWindow.cpp
Increased the maximum size of the block of selected samples in Plot Spectrum. This now enables 46.60 minutes of a 48kHz sampled signal to be analyzed. The previous limit often prevented the analysis of a complete song. The dynamically allocated memory for storing the selection is now being freed when the Plot Spectrum dialog is exited. This is better housekeeping, which is more important now that we can have a much larger selection.