Remove brush button from Tools toolbar, and...
... Remove it too from Extra > Tools; and:

Make the rotation of tools with A and D keys include the brush in the
dialog, and make presses of any of these buttons pop up the others.

(Go to Preferences > Keyboard, press Defaults, choose Full, to enable the A and
D keystrokes.  Or go to Preferences > Interface, enable extra menus, and see
Extra > Tools > Previous/Next Tool.)

Also don't hide the dialog for change of tool