Implement snapping of the moving end of changing play region...
... With Esc key too, to cancel the snap without cancelling the drag.

So you can snap to the start or end of a clip or label.

But still not implemented (as for changes of selection) is snapping of the free
end of the region before left button down.

It also might be useful to snap to the start or end point of the selection.