issue #1780: An unwanted label can be created by typing
If a label is selected, then you can type to create a label, even if the "Type to create a label" option is disabled.

This bug was introduced in commit: a0ad72d
I presume that in the following line in LabelTrackView::DoCaptureKey:
if (IsValidIndex(mTextEditIndex, project) || IsValidIndex(mNavigationIndex, project)) {,

IsValidIndex(mNavigationIndex, project) was included to let the hardcoded Enter key be captured. This was not an appropriate way of doing so - the Enter key should have been handled individually, just like the tab key. This introduced the current bug.

Fix: Given that use of the hardcoded Enter key is shortly to be removed, then IsValidIndex(mNavigationIndex, project) can be removed.