Less laggy response of looping playback to change of region bounds...
... Some lagginess has to be tolerated, because there must be a queue of some
length between the TrackBufferExchange thread and the PortAudio thread.   But
how short can it be?

In case of playback looping, that lag is now set to an arbitrary hard-coded
1/2 second in LoopingPlaybackPolicy::SuggestedBufferTimes; which is less than
the old default latency of 4 seconds (see PlaybackPolicy::SuggestedBufferTimes),
which perhaps was a value motivated by older generations of hardware.

So, if you adjust looping play bounds while the play head is within 1/2 second
of the end of the loop (that's 1/2 second of real time, not track time, in case
you also have a time-warping Time Track) -- then you might yet notice the play
head not doing quite what you want.

(Should we make that latency user-tunable?  Or, more challengingly for us,
detect hardware limitations and make it adaptive?)