Don't make black corners of control toolbar buttons when blending...
... Though this isn't the only consequence of unwary use of ThemePrefs.  It is
only the more obvious one.

Actually, the bitmaps made by OverlayImage in ImageManipulation.cpp, which
combines a foreground and a background image, are opaque.  But the
transparencies in the given background image are blended with the theme's
clrMedium which is the background color of the toolbar.  This change makes
sure that the background image does not lose its transparencies when it is
recolored.