Allow for the playhead time to be less than left loop bound...
... This will matter where NewDefaultPlaybackPolicy::MessageConsumer calls
RealTimeInit.

(This also makes backwards seeking unbounded (by the old method of seeking that
should be phased out in future).  See the few uses of AudioIO::SeekStream().
These would be better handled with messages to playback policies on the
producer side -- not the draining of the ring buffers, then restarting of the
producer, all done by the consumer.)