Attempt to fix UB in StopStream
wxYield causes the event loop to continue processing events, while not all changes were applied to the tracks. In this particular case  AffordanceHandle was created with the old track from TrackList(during wxYield call), which existed before recording has started, then AffordanceHandle::Release was called when the track is already removed from the TrackList and replaced with a pending one.

Though the consequence is that without wxYield the UI thread is completely blocked until the AudioCallback thread signals it's done. Definitely needs testing.