Fix violation of C++ ODR rule
The anonymous linkage of these three classes meant that when wrapped
inside a weak_ptr, they would not be the same object when the header was
instantiated in multiple locations. This violated the C++ one definition
rule. The fix is to remove the anonymous namespace and give these types
external linkage.