Fix Issue #2239: crash cancelling Nyquist prompt...
... To reproduce:

* Generate a tone
* Open the Nyquist Prompt and enter the code:

```
;control test "Test" float "" 5 0 10
(print test)
```

* Click the "OK" button
* Click the "Cancel" button
* Observe crash

Bug was introduced at b1e988e4f50ff19050d00c8c5685d43a9fc8a62b
The nulling-out of a pointer to the Effect in the Close routine was deleted,
but the destructor is still called later.

To fix it, don't unnecessarily repeat in the destructor of the modal dialog
(which may be delayed until an event loop) what was already done when closing
it.