Fixes an unexpected performance regression in SpectrumView
After merging the spectrum brush branch - a new hot path was accidentally added to spectrum rendering.

1. std::round is expensive as is. The implementation is complex and designed to handle different situations.
2. By mistake the following chain was introduced: round(float)->int->double->round(double)->int

The new implementation uses `lrintf`. By default, it rounds to nearest, and ties are rounded to even. This does not match the behavior `std::round` entirely, but given that it only slightly affects the visuals, that shouldn't be a problem