Fixes the behavior of Locate FFmpeg dialog. (Bug #2282)
The meaning of the "Location of " field became confusing after the support for multiple FFmpeg versions was added. 

This commit changes the way how the content of the dialog is interpreted. 
Both full paths to avformat and to the directory containing avformat are now valid.

Additionally, it fixes the way Audacity checks that FFmpeg works. Audacity will check if it can load the library instead of checking that the file exists.