issue #2078: tabbing does not ensure selected label is visible
Problem:
Tabbing used to open the name of the selected label for editing. When this was the case, LabelTrackView::KeyDown() ensured that cursor was visible.
However, this was changed so the the label is selected, and not automatically opened for editing.

Fix:
Add a call to scroll so that the start of the selected label is visible whenever tab or shift + tab is pressed.