Restrictions on the view of wxWidgets through the wxBase target...
... wxBase is no longer an alias target.  It is an interface target to wxWidgets
that adds some preprocessor definitions that restrict at compile time what
parts of wxWidgets may be used, to general wxBase utilities only.

No graphical interface is allowed, but also none of the main event loop.  So in
fact only a subset of wxBase is good to use, because even that includes an event
loop usable in console programs.  Restrict the use of certain headers by
pre-including their include guards, so that an attempt to include them skips
the definitions they contain and makes compilation fail if you try to use them.

The library targets that depend on wxBase must make it PRIVATE, for otherwise
the restrictions on use of wxWidgets would be propagated to users of those
libraries that do not need to be toolkit-neutral and break their compilation.

Restoring what was defined before at 1f71ef4 but was soon lost at f66381b with
the conan-izing.