Issue2372: Nyquist can hang ...
... Problem was xlbegin() calls not balanced by xlend().  This can cause
the global variable xlcontext in the XLisp interpreter to be a dangling pointer
to an expired stack variable.  When re-entering the interpreter, the context
was again constructed at the coincidental same address, but it should have been
NULL.  This caused xlreturn() to enter a tight infinite loop, when evaluating
a simple `(return 0)` in Nyquist prompt.

I reviewed and fixed all unbalanced xlbegin() calls, other than in xlisp_main().

Those in nyx.c affect Audacity, one of them when evaluating a sound, another
when evaluating other values; but should not be reached in standalone Nyquist.

Also: fix dangling else in xlread.c in case DEBUG_INPUT is enabled.