Temp interface reenables addition of one effect to global stack...
... This is a non-permanent expedient for testing.

Undo and redo of changes to effect stacks might be implemented, and should be
done at the level of EffectUI.cpp, not in RealtimeEffectManager.cpp where it
would make undesirable dependency on UndoManager.

However, the undo and redo of RealtimeEffectList objects still awaits
implementation, and undo history pushes would not be desirable at the places
where state adding and deleting are inserted in this commit -- where a single
non-modal effect dialog is opened and closed.