Require an EffectSettingsAccess to build effect dialogs...
... Passing by reference is the way to say "null not allowed!"

For now the settings are supplied in two places from the default settings
saved in EffectManager, for destructive effect application.  In future there
will be other EffectSettings objects stored elsewhere, such as per-track in
effect stacks for nondestructive processing.

Intermediate levels of the call stack are a little confusing.

But ultimately it is the EffectUIHost dialog, which may be non-modal, that
extends the lifetime of the access object, by converting it to a shared_ptr
with shared_from_this()