Define an EffectSettingsAccess class for RealtimeEffectState...
... This inserts the calls to CopySettingsContents, so that the worker thread
avoids free store allocations, and it performs a proper atomic transfer of
whatever is in EffectSettings.

There are two MessageBuffers, carrying changes of settings in two directions
between threads.  CopySettingsContents must happen during an assignment when
writing a MessageBuffer slot, and during a constructor when reading, so that
the synchronization in MessageBuffer works correctly:  all must happen
between the two atomic exchanges.  This explains the template arguments of the
two MessageBuffers.