Transportstate fix - part 1 (#2554)
* Move mpTransportState after request to process "once"

Alone this simple moving resolved the issue, because:
in all instances where mpTransportState was found to be empty during ::TransformBuffers, the call stack
showed that the call was originated in response to a request to process "once", and never in response to process "LoopRunning".

In ::StopStream, a call to process "once" is always requested, because `if (mStreamToken > 0)` will always
be satisfied - I can not seem to reproduce a case when it is not (but somebody else might know better!)

As long as `if (mStreamToken > 0)` always happens, there is no point in introducing what we discussed earlier,
(i.e. a new atomic and a waiting loop on that in order to signal that "LoopRunning" processing is really stopped)
it would be an unnecessary complication, because as the atomic to process "once" is later set to true and then
waited for until it is "false", this will always happen after the last processing coming from "LoopRunning".

However, if there are ever cases when `if (mStreamToken > 0)` is false, then we need this extra atomic, to wait
to be sure that "LoopRunning" processing came to a stop, before resetting mpTransportState. I have all this extra
things ready into a stash anyway - first we must decide if this extra complication is necessary.

* reset mpTransportState closer to after the last process + stronger memory ordering

* stricter memory order applied in other places

Co-authored-by: Pietro Marcello <Pietro.Marcello@xperi.com>
Co-authored-by: pietro68 <p.marcello@mu.se>