Rewrite smart pointer for Apple Core resources, use more widely...
... Not that AudioUnitGetProperty and AudioUnitSetProperty take void* and size
arguments, and they can simply reinterpret the smart pointer as just a pointer.

This safely seats pointers that require releases directly into the smart pointer
when getting.

Note the convention in SDK functions, that "Create" and "Copy" functions return
pointers that require the user to Release them, but "Get" functions do not
(unless the user "Retains" them, increasing reference count).

The comment in SDK for TISGetInputSourceProperty says that.