Remove old no-argument overload of Effect::Startup() ...
... Consequence will only be, that if some user migrates from Audacity 2.1.0 or
earlier directly to 3.2.0 or later -- then, their default settings for the
built-in effects will not be preserved in the preferences, but rather all will
go back to the built-in defaults.

It is unlikely anyone will do that.

This simplification eliminates some uses of the state variables that are in
the built-in effect classes, but ought to be externalized as EffectSettings.
Rewriting all these functions with an EffectSettings argument isn't worth the
trouble.