Pass EffectSettings to even more places...
... So that we can smuggle the duration into ProcessInitialize() where some
effects will need it.

But to keep the virtual functions uniform for all effects, we have to pass it
also to RealtimeAddProcessor which calls ProcessInitialize().

There is one doubtful place in VST2 effects where this is bypassed.  It has no
analogue for the other third party effect families.  It will not be important,
for reasons stated in the comments.