EffectUIHost always calls TransferDataToWindow after settings changes...
... This is a virtual function of dialogs that updates UI controls from stored
values, and some effects (which are not subclasses of wxDialog) implement its
steps with another virtual function with the same name.

(What should happen in future is elimination of this virtual function of Effect.
Instead, wxValidator should be used with all controls.  That will make the
default implementation of wxDialog::TransferDataToWindow() sufficient.  This
work may not be a very far future but it is large.)

Don't do it in Effect::SetAutomationParametersFromString or from built-in
effect subclasses.

Also remove an unused function in Distortion.

Make the back-pointer to the dialog private!  Effect sub-classes should not
need awareness of whether there is a dialog present.