Rewrite two unusual uses of TransferDataToWindow...
... These lie on paths where EffectSettings will not be available but won't
need EffectSettings.

In Amplify, there is some adjustment of the ratio and other state variables
that depend on it.  This should not happen when transferring to the window,
but from it.

In AutoDuck, there is a special panel that needs to update some values in
response to mouse movements and refresh itself.  This can ultimately be
rewritten as a function of an EffectUIValidator which will store a reference to
EffectSettingsAccess.  For now just call through to a low level function
DoTransferDataToWindow, leaving TransferDataToWindow free to take
a new EffectSettings & argument.