EffectUIHostDialog::TransferDataFromWindow calls-through to base...
... And also first invokes Validate() in the base.

And it afterward invokes whatever remains in the ValidateUI() member of
EffectUIValidator and the TransferDataFromWindow() override of the effect.

This causes any wxValidator objects attached to controls to check and transfer
values.

Supplying this long-missing call-through allows simplifications of other data
transfer functions.

Also rename the member function of EffectUIValidator to ValidateUI().