LoadUserPreset takes Settings and is const...
... Again, cheat with const_cast in many stateful effects.  Some const_cast
added in Effect, but one const_cast can be removed.

Comments explain better how it cooperates with SaveUserPreset in the case of
built-in effects.

The transitional DUMMY in Effect::SetAutomationParametersFromString is used
again.  Notice there is mutual recursion between that and
Effect::LoadUserPreset, and that it too becomes const.

The previously added, and explained, dummy settings object in NyquistEffect.cpp
is used again.  A dummy is added in Effect::UnsetBatchProcessing which is
like the one in Effect::SetBatchProcessing.  These dummies do not require later
fixes.

Comments in EffectManager::SetEffectParameters and related.

SyncParameters() in VST3 now is const and takes EffectSettings & which it may
mutate when its state is externalized in future.