AudacityCommand has two virtual functions for visiting settings...
... one taking a SettingsVisitor, the other a ConstSettingsVisitor.

This is a big repetitious commit, but note in some places that types of
arguments needed to be changed so overloading resolution of Define() would work
correctly.

Maybe in future some template cleverness (as with effects) could generate these
functions, and also generate the PopulateOrExchange overrides of commands,
just from a compile-time list of controls?