No more Effect::mClient ! ...
... Third party effects don't need a "host" object of class Effect to forward
messages to them.  Instead they implement the same EffectUIHostInterface,
although that is still a messy thing.

Many recent commits worked toward this.  First the confusing
EffectHostInterface, which was supplied to EffectUIClientInterface for
initialization -- had to be eliminated.

That might have been done more simply just by having the other classes
inherit Effect sooner.  But taking the long path that we did
discovered the need for duration to be treated as part of
EffectSettings, because it was part of persistent settings.

To do that, required the passing of EffectSettings to many more places --
which would also have been needed ultimately, to allow built-in
effects to become stateless.

Note: without mClient, Effect::Startup would do nothing but pass through the
return code of InitializeInstance.  So just call that directly in
EffectManager.  That virtual function is moved to a different abstract base
class.

Also eliminate the Startup() in EffectManger::NewEffect.  Return a pointer (no
longer a unique_ptr) to the effect, which RealtimeEffectState should treat as
a stateless object that is a factory for state.  But -- that's not really what
happens, yet.  To make it so is matter for yet OTHER long pull requests!