Rewrite the saving and loading of the last used duration...
... Save it from the Effect UI dialog when transferring from the window.  Most
of the existing uses of SetDuration(), where it used to happen, happen while the
dialog is open; this updates the configuration in just one common place when
the dialog is being closed, without cancellation.

So don't do it when SetDuration() is used in Effect::DoEffect.

Don't repeat the string literal for the registry key.