EffectDtmf externalizes its settings ...
... But still inherits StatefulPerTrackEffect

Notes:

dtmfNTones is now size_t not int.

Parameters.Reset() has to be removed because it supplies a dummy EffectSettings
and would now violate the assertion in GetSettings().

The Effect object no longer holds settings, but the Validator holds its own
complete settings -- which have a short life (only while a dialog is open).

UpdateUI() uses the EffectSettingsAccess to update the validator's copy of
settings in case of any changes communicated from outside.

ValidateUI() and event handlers use the access to send settings changes back to
the framework.