DtmfEffect no longer inherits StatefulPerTrackEffect
Note:

Effect::mT0 is STILL a piece of state in the base class Effect that should be
moved to EffectContext instead.  (That class does not yet exist.)  Its value
is copied to the validator when it is constructed.

Also the use of Effect::MessageBox depends on a pointer to the dialog, but
only for the minor purpose of choosing the parent of the dialog box.

See two TODO comments.