Move default implementation of Effect::Process to new subclass...
... PerTrackEffect, an abstract subclass of Effect.

It makes Process() final.  Effects, that did not need to override this further,
implemented their special work mostly with ProcessBlock(), and some other
overrides that setup and clean up for that.

TwoPassSimpleMono, used only by Compressor, is now unrelated to this extension.

The variable holding block size in the base class effect is renamed, to be sure
it is not used for anything else in Effects that do not inherit PerTrackEffect.
Indeed, the new member name was only needed in the get and set functions.

ProcessPass and ProcessTrack do not need to be virtual