Untangle control flow for recursive Nyquist effect...
... When the Nyquist prompt is given a script with magic ;control comments,
which cause a second Nyquist effect and dialog.

Don't do the effect during ShowHostInterface() -- do it in Process(), so that
the Delegate() is like other calls to Delegate().  This allows Effect::Delegate
to simplify with fewer arguments.

Example input to the Nyquist prompt for testing this -- a simple ad-hoc amplify:

;control slider “slider” real “” 0.5 0 1
(mult slider *track*)