Change the representation of track groups...
... It used to be just a boolean, set to true only in left channels of stereo
tracks.

In 3.1 the boolean became an enumeration, to carry some extra information.

Now, it becomes a unique pointer to a structure that can carry still more
information that applies to the group of channels, rather than to one channel
alone.

A track may hold non-trivial group information but also be the sole member of
its group.

Track::DoSetLinkType() does all the work of keeping the track list consistent.
When the track is the first of its group, carrying the group data, then any
partner track must have null group data.  When a group is split, group data is
replicated.

Generalization beyond two channels maximum isn't yet implemented, but the
details of such a future development could be encapsulated in Track.cpp