Simplify calls to get AudioUnit properties of fixed size...
... This is similar to what we just did for AudioUnitSetProperty.  But not all
calls to AudioUnitGetProperty are rewritten.

There are some where there is first a query of object size, using
AudioUnitGetPropertyInfo.  Then there is an allocation of space, and then a
call to AudioUnitGetProperty.

Those calls will be simplified in the next commits.

Note that AudioUnitGetProperty can retrieve less of the data than are available
from the plug-in if it doesn't first check with AudioUnitGetPropertyInfo and
it passes a small size.

The SDK appears to do this to allow versioning of the structures while not
breaking code written to prior versions.

In some cases, as mentioned in some comments, there is an array of data
available yet only the first element of that array is of interest to us.