Statelessness for AudioUnits...
... Not using the template EffectWithSettings, because we define all the names
it defines, differently.

Note the correction needed in RealtimeProcessStart, to complete the inter-thread
communication of settings changes during play.

MakeSettings and CopySettingsContents are appropriately defined.

CopySettingsContents has some important assertions.  Fetching the parameter
list only once at plug-in initialization, then iterating that same list
completely in MakeSettings, guarantees those assertions -- provided that no
other code ever inserts or removes keys in the map.

The container changes from unordered_map to map, so that the iteration order is
determined in CopySettingsContents.

GetSettings is explicitly defined, just as EffectWithSettings would generate it.
But it is put in the class AudioUnitWrapper so AudioUnitInstance can use it too.