Define StoreSettings, complementary to FetchSettings...
... And use it in SavePreset which complements LoadPreset, and in Export which
complements Import.

And review the uses of SavePreset.  One really implements the saving of custom
user presets.

Another is for the one-time dumping of the initial plugin setting state to the
configuration file, to be used afterward in LoadFactoryDefaults.

The remaining use of SavePreset is only in migration of old (likely only very
old) configuration files.  Perhaps that logic can be removed.