Fix some confusions about instance vs plug-in in AudioUnits...
... See explanation in comment in InitializePlugin().  The AU SDK is itself a
source of some confusion, compared with other SDKs.

See commit 7a12a2e where that function name disappeared but now it is brought
back.

RealtimeAddProcessor now calls a really boolean valued function, not
creating and destroying the so-called Instance object (which still has only a
stub implementation, StatefulPerTrackEffect::Instance).

Some things are now done in InitializePlugin(), which is called in two places,
but not also in InitializeInstance().  Nothing changed near those two calls.

Note in this and following commits that many SDK functions (like
AudioComponentInstanceNew) return 0 for success and non-zero for failure, but
we never needed to distinguish different non-zero return values.