Pass EffectInstance into Effect::PopulateOrExchange overrides...
... anticipating the needs of AudioUnit, VST, and VST3 effects, to communicate
changes of controls to the plug-ins.  (Ladspa and LV2 don't need this because
with them, it is the host program that decides where parameter values are in
memory.)

Do all else needed to pass an instance through intermediate levels of stack,
starting at EffectBase::DoEffect().

Make dummy instances in EffectManager and Nyquist, so that ShowHostInterface
can take the instance by reference instead of a possibly null pointer.

Nothing is done yet with the argument.