Rewrite track link consistency checking...
... Make a non-mutating version that only checks without doing repairs, safe
to use in a debug assertion.

Don't violate the assertion incorrectly when the reloading of the list of tracks
is in progress.

Fixes error from 93be3aeb that could crash debug build when opening a project.

All other uses of LinkConsistencyCheck() predated that commit, and are changed
to call the new name LinkConsistencyFix().