FileConfig is non-abstract...
... Its Warn() override does nothing, instead of putting up a dialog to indicate
failure of initialization of the config file because it is read-only.

Intended for use in a worker process that will not fully initialize a wxWidgets
application, so can't show a window.

There will always be a main process that detects the same problem and can tell
the user about it.