Some C++ wrappers for structs in the AudioUnits sdk...
... Mostly valuable for some defaulting of some members to common values, as
in UserPreset, or for defining a destructor.  (Also moving smart pointer
template AudioUnitCleaner)

This is probably unnecessarily cautious for future-proofing the code.

They are not necessary to allow use of the concise aggregate initialization
syntax with { }, or to get zero-initialization of extra fields in a future
version of the structure.  (Refer to aggregate initialization details for C++.)

If fields happen to get reordered, then they would still be correctly
type-checked and reassigned.  But I doubt Apple would make such foolish ABI
breaking changes, breaking applications porting to later operating system
versions!

(FFmpeg was foolish that way, though... a different story Dmity can tell you
about.)